refactor the local/remove proxy DB intializing
Also the HTTP listener was removed by mistake
