+++ cmd/telliot/main.go
@@ -108,26 +108,6 @@ func migrateAndOpenDB(cfg *config.Config) (db.DB, error) {
 	return DB, nil
 }
 
-func createProxy(cfg *config.Config, DB db.DB) (db.DataServerProxy, error) {
-	var dataProxy db.DataServerProxy
-	if cfg.Mine.RemoteDBHost != "" {
-		proxy, err := db.OpenRemoteDB(cfg, DB)
-		if err != nil {
-			return nil, errors.Wrapf(err, "open remote DB instance")
-
-		}
-		dataProxy = proxy
-	} else {
-		proxy, err := db.OpenLocalProxy(DB)
-		if err != nil {
-			return nil, errors.Wrapf(err, "opening local DB instance:")
-
-		}
-		dataProxy = proxy
-	}
-	return dataProxy, nil
-}
-
 var cli struct {
 	Transfer transferCmd `cmd:"" help:"Transfer tokens"`
 	Approve  approveCmd  `cmd:"" help:"Approve tokens"`
