+++ docs/setup-and-usage.md
@@ -56,9 +56,13 @@ cp configs/config.json .local/configs/$NAME/config.json # Edit the file after th
 cp configs/indexes.json .local/configs/$NAME/indexes.json
 cp configs/manualData.json .local/configs/$NAME/manualData.json
 # Add the configs.
-kubectl create configmap telliot-$NAME --from-file=.local/configs/$NAME/config.json  --from-file=.local/configs/$NAME/indexes.json --from-file=.local/configs/$NAME/manualData.json -o yaml --dry-run=client | kubectl apply -f -
+kubectl create configmap telliot-$NAME \
+  --from-file=.local/configs/$NAME/config.json \
+  --from-file=.local/configs/$NAME/indexes.json \
+  --from-file=.local/configs/$NAME/manualData.json \
+  -o yaml --dry-run=client | kubectl apply -f -
 
-# Copy the deployment and create it.
+# Copy the deployment and run it.
 cp configs/manifests/telliot.yml .local/configs/$NAME/telliot.yml
 sed -i "s/telliot-main/telliot-$NAME/g" .local/configs/$NAME/telliot.yml
 kubectl apply -f .local/configs/$NAME/telliot.yml
