+++ pkg/ops/dataServerOps.go
@@ -33,8 +33,7 @@ func CreateDataServerOps(
 	ctx context.Context,
 	logger log.Logger,
 	config *config.Config,
-	DB db.DB,
-	proxy *db.DataServerProxy,
+	DB db.DataServerProxy,
 	client rpc.ETHClient,
 	contract *contracts.Tellor,
 	account *rpc.Account,
