+++ pkg/tracker/disputeStatus_test.go
@@ -23,8 +23,10 @@ func TestDisputeString(t *testing.T) {
 	cfg := config.OpenTestConfig(t)
 	DB, cleanup := db.OpenTestDB(t)
 	defer t.Cleanup(cleanup)
+	proxy, err := db.OpenLocal(cfg, DB)
+	testutil.Ok(t, err)
 	logger := util.SetupLogger("debug")
-	tracker := NewDisputeTracker(logger, cfg, DB, nil, nil)
+	tracker := NewDisputeTracker(logger, cfg, proxy, nil, nil)
 	res := tracker.String()
 	testutil.Assert(t, res == DisputeTrackerName, "didn't return expected string", DisputeTrackerName)
 }
@@ -38,20 +40,22 @@ func TestDisputeStatus(t *testing.T) {
 
 	DB, cleanup := db.OpenTestDB(t)
 	defer t.Cleanup(cleanup)
+	proxy, err := db.OpenLocal(cfg, DB)
+	testutil.Ok(t, err)
 	logger := util.SetupLogger("debug")
 	contract, err := contracts.NewTellor(cfg, client)
 	testutil.Ok(t, err)
 	account, err := rpc.NewAccount(cfg)
 	testutil.Ok(t, err)
-	tracker := NewDisputeTracker(logger, cfg, DB, &contract, &account)
+	tracker := NewDisputeTracker(logger, cfg, proxy, &contract, &account)
 	testutil.Ok(t, tracker.Exec(context.Background()))
-	v, err := DB.Get(db.DisputeStatusKey)
+	v, err := proxy.Get(db.DisputeStatusKey)
 	testutil.Ok(t, err)
 	b, err := hexutil.DecodeBig(string(v))
 	testutil.Ok(t, err)
 	t.Logf("Dispute Status stored: %v\n", string(v))
 	testutil.Equals(t, b.Cmp(big.NewInt(1)), 0, "dispute status from client did not match what should have been stored in DB. %s != %s", b, "one")
-	DB.Close()
+	proxy.Close()
 }
 
 func TestDisputeStatusNegativeBalance(t *testing.T) {
@@ -63,14 +67,16 @@ func TestDisputeStatusNegativeBalance(t *testing.T) {
 
 	DB, cleanup := db.OpenTestDB(t)
 	defer t.Cleanup(cleanup)
+	proxy, err := db.OpenLocal(cfg, DB)
+	testutil.Ok(t, err)
 	logger := util.SetupLogger("debug")
 	contract, err := contracts.NewTellor(cfg, client)
 	testutil.Ok(t, err)
 	account, err := rpc.NewAccount(cfg)
 	testutil.Ok(t, err)
-	tracker := NewDisputeTracker(logger, cfg, DB, &contract, &account)
+	tracker := NewDisputeTracker(logger, cfg, proxy, &contract, &account)
 	testutil.Ok(t, tracker.Exec(context.Background()))
-	v, err := DB.Get(db.DisputeStatusKey)
+	v, err := proxy.Get(db.DisputeStatusKey)
 	testutil.Ok(t, err)
 	b, err := hexutil.DecodeBig(string(v))
 	testutil.Ok(t, err)
