+++ pkg/tracker/tributeBalance_test.go
@@ -28,15 +28,17 @@ func TestTributeBalance(t *testing.T) {
 
 	DB, cleanup := db.OpenTestDB(t)
 	defer t.Cleanup(cleanup)
+	proxy, err := db.OpenLocal(cfg, DB)
+	testutil.Ok(t, err)
 	logger := util.SetupLogger("debug")
 	contract, err := contracts.NewTellor(cfg, client)
 	testutil.Ok(t, err)
 	account, err := rpc.NewAccount(cfg)
 	testutil.Ok(t, err)
-	tracker := NewTributeTracker(logger, DB, &contract, &account)
+	tracker := NewTributeTracker(logger, proxy, &contract, &account)
 	err = tracker.Exec(context.Background())
 	testutil.Ok(t, err)
-	v, err := DB.Get(db.TributeBalanceKey)
+	v, err := proxy.Get(db.TributeBalanceKey)
 	testutil.Ok(t, err)
 	b, err := hexutil.DecodeBig(string(v))
 	testutil.Ok(t, err)
